init_repository_fix1
