+++ app/src/main/java/com/example/nguyenthanhtungh/moviedb/data/repository/GenreRepository.kt
@@ -1,13 +1,11 @@
 package com.example.nguyenthanhtungh.moviedb.data.repository
 
 import com.example.nguyenthanhtungh.moviedb.data.source.GenreDataSource
-import com.example.nguyenthanhtungh.moviedb.data.source.local.GenreLocalDataSource
-import com.example.nguyenthanhtungh.moviedb.data.source.remote.GenreRemoteDataSource
 import com.example.nguyenthanhtungh.moviedb.data.source.remote.response.GenreResponse
 import io.reactivex.Single
 
-class GenreRepository(private val local: GenreLocalDataSource,
-                      private val remote: GenreRemoteDataSource)
+class GenreRepository(private val local: GenreDataSource.LocalDataSource,
+                      private val remote: GenreDataSource.RemoteDataSource)
     : GenreDataSource.LocalDataSource, GenreDataSource.RemoteDataSource {
 
     override fun getListGenres(): Single<GenreResponse> {
