+++ app/src/main/res/layout/genre_item.xml
@@ -13,12 +13,12 @@
         <TextView
             android:id="@+id/genre_item_name"
             android:layout_width="0dp"
-            android:layout_height="25dp"
-            android:layout_marginStart="8dp"
-            android:layout_marginLeft="8dp"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="8dp"
-            android:layout_marginRight="8dp"
+            android:layout_height="@dimen/dp_25"
+            android:layout_marginStart="@dimen/dp_8"
+            android:layout_marginLeft="@dimen/dp_8"
+            android:layout_marginTop="@dimen/dp_8"
+            android:layout_marginEnd="@dimen/dp_8"
+            android:layout_marginRight="@dimen/dp_8"
             android:autoSizeMaxTextSize="@dimen/sp_24"
             android:autoSizeMinTextSize="@dimen/sp_12"
             android:autoSizeTextType="uniform"
